+++ language-server/Shared.AnZwDev.ALTools.Server/Contracts/CodeCompletionRequest.cs
@@ -1,4 +1,5 @@
 ﻿using AnZwDev.ALTools.ALSymbols;
+using AnZwDev.ALTools.CodeCompletion;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -11,6 +12,7 @@ namespace AnZwDev.ALTools.Server.Contracts
         public Position position { get; set; }
         public string path { get; set; }
         public List<string> providers { get; set; }
+        public CodeCompletionParameters parameters { get; set; }
 
     }
 }