+++ vscode-extension/src/langserver/codeCompletion/toolsCodeCompletionRequest.ts
@@ -1,15 +1,18 @@
 import { TextPosition } from "../../symbollibraries/textPosition";
+import { ToolsCodeCompletionParameters } from "./toolsCodeCompletionParameters";
 
 export class ToolsCodeCompletionRequest {
     
     position: TextPosition;
-    path: string;
+    path: string;    
     providers: string[];
+    parameters: ToolsCodeCompletionParameters;
 
-    constructor(newPosition: TextPosition, newPath: string, newProviders: string[]) {
+    constructor(newPosition: TextPosition, newPath: string, newProviders: string[], newParameters: ToolsCodeCompletionParameters) {
         this.position = newPosition;
         this.path = newPath;
         this.providers = newProviders;
+        this.parameters = newParameters;
     }
 
 }